ft_strnstr to ft_strdup

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: aoberon <aoberon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 16:06:42 by aoberon           #+#    #+#             */
-/*   Updated: 2022/11/08 16:13:20 by aoberon          ###   ########.fr       */
+/*   Updated: 2022/11/10 10:20:58 by aoberon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *restrict dest, const void *restrict src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i != n)
